* wheel event

+++ src/Mouse.cpp
@@ -4,7 +4,7 @@ bool Mouse::isButtonPressed(Button button) {
     return glfwGetMouseButton(glfwGetCurrentContext(), button) == GLFW_PRESS;
 }
 
-glm::vec2 Mouse::getPosition() {
+glm::dvec2 Mouse::getPosition() {
     glm::dvec2 pos;
     glfwGetCursorPos(glfwGetCurrentContext(), &pos.x, &pos.y);
     return pos;
